+++ example/src/App.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import { UseDebounce } from './UseDebounce'
 import { UseEventListener } from './UseEventListener'
 import { UseInput } from './UseInput'
 import { UseLocalStorage } from './UseLocalStorage'
@@ -7,9 +8,10 @@ import { UseLockBodyScroll } from './UseLockBodyScroll'
 const App = () => {
   return (
     <>
+      <UseDebounce />
+      <UseEventListener />
       <UseInput />
       <UseLocalStorage />
-      <UseEventListener />
       <UseLockBodyScroll />
     </>
   )