+++ src/index.js
@@ -1,3 +1,4 @@
+export * from './useDebounce'
 export * from './useEventListener'
 export * from './useInput'
 export * from './useLocalStorage'